Move hash type declarations to hashing.hpp
